requirements.txt
Flask==0.10.1
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Pillow==2.9.0
python-telegram-bot==2.8.5
requests==2.7.0
Werkzeug==0.10.4
